Fix Vue wrappers missing form control props for Checkbox and Radio
The VueCheckbox and VueRadio wrappers were missing the new form control
props (label, labelHidden, noLabel, required, invalid, errorMessage, helpText)
that were added during the form control refactoring.

This caused props not to be passed through to the web components, resulting
in attributes not appearing in the DOM and features not working properly.

Changes:
- VueCheckbox: Added all 7 form control props to interface and withDefaults
- VueCheckbox: Added prop bindings in template
- VueRadio: Added all 7 form control props to interface and withDefaults
- VueRadio: Added prop bindings in template

Verified:
- Input, Select, Combobox, Slider, Toggle already had correct props
- All component tests passing